requirements.txt
Flask
pydantic
python-dotenv
openai
google-api-python-client
Pillow
pytest
